% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raneff.R
\name{get_raneff}
\alias{get_raneff}
\title{Tidy random effect estimates from INLA}
\usage{
get_raneff(inla_mod, scales = c("var", "sd"))
}
\arguments{
\item{inla_mod}{A fitted INLA model}

\item{scales}{Either "var", or "sd".}
}
\value{
Tibble with random effect estimates.
}
\description{
Tidy random effect estimates from INLA
}
