% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raneff.R
\name{hyper_sd}
\alias{hyper_sd}
\title{Transform precision to sd or var.}
\usage{
hyper_sd(prec, scales, internal = FALSE)
}
\arguments{
\item{prec}{Precision}

\item{scales}{"var" or "sd"}

\item{internal}{}
}
\value{
transformed vectors
}
\description{
Transform precision to sd or var.
}
